fix(core): 解决数据库服务依赖和空值处理问题
- 在archive_service中添加数据库服务依赖检查和空值验证
- 修复asset_info_view中的空值处理，防止None引用错误
- 在多个视图组件中添加asset为空的保护性检查
- 优化card_automation_service中模板为空的异常处理
- 修复media_import_service中资产为空时的处理逻辑
- 在raw_extraction_service中添加目标路径为空的安全检查
- 修复project_dashboard_view中项目ID为空的处理
- 在report_service中改进类型注解和数据安全处理
- 修复restore_wizard中日志记录格式问题
- 添加pyright配置文件以增强类型检查支持

+++ DITWorkstation/DITWorkstation/main.py
@@ -132,7 +132,7 @@ def _create_app_icon() -> QIcon:
     font = QFont("Helvetica", int(size * 0.6), QFont.Bold)
     painter.setFont(font)
     painter.setPen(QColor("#ffffff"))
-    painter.drawText(QRect(0, -size * 0.03, size, size), Qt.AlignCenter, "D")
+    painter.drawText(QRect(0, int(-size * 0.03), size, size), Qt.AlignCenter, "D")
     painter.end()
     return QIcon(pixmap)
